+++ CHANGELOG.md
@@ -7,7 +7,7 @@ Note that since we don't clearly distinguish between a public and private interf
 ## [Unreleased]
 
 - Fix measurement label `offsetZ` default: not needed when `scaleByRadius` is enabled
-- ModelServer SDF/MOL2 ligand export: fix atom indices when additional hydrogen atoms are present
+- ModelServer SDF/MOL2 ligand export: fix atom indices when additional atoms are present
 
 ## [v3.43.1] - 2023-12-04